Rename reproduction to TestIndexer.test.ts as a general TestIndexer suite
Keeps the cross-batch cached-effect load case but drops the repro-specific
naming so future TestIndexer cases can live in the same file.
